Ahaggar: make buffer size default to 4096
The default 1024 size is triggering a memory corruption when analyzing
some large code.

During some rare cases the xrealloc() call will fail, glibc free() calls
will detect that the heap and pointers were corrupted, and it will produce
a crash dump.

Currently I am not sure if this bug is in the current code, or due to
some external factors which I don't have time to investigate.
However:
 it seems that this is not a use-after-free bug since making the buffer
 size to 4096 will avoid any corruption.

So Just make the default buffer size to 4096 instead of 1024. Note that
the buffer will increase sometimes automatically to 24576 when analyzing
large code without any corruption, so... next time!

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>