Ahaggar: Add the fncalls-match.c file
This file contains the definition of the routines used to manipulate
internal hash tables that are used to store functions and pointers.

Currently for functions matching we use static arrays by just including
elements from external header files.
We can later switch to dynamic data structures to be more flexible.

Signed-off-by: Djalal Harouni <tixxdz@opendz.org>